Refactor Carousel/Banner
