+++ data/offerings.js
@@ -1,0 +1,77 @@
+export const offerings = [
+  {
+    Name: 'Workshop',
+    Caption: 'Immersive Learning!',
+    contentPosition: 'left',
+    Items: [
+      {
+        Name: 'Workshop',
+        Image: 'https://source.unsplash.com/featured/?presenter',
+      },
+      {
+        Name: 'Classroom',
+        Image: '/images/Workshop.png',
+      },
+    ],
+  },
+  {
+    Name: 'Webinar',
+    Caption: 'Virtual Conference!',
+    contentPosition: 'middle',
+    Items: [
+      {
+        Name: 'Zoom',
+        Image: 'images/Webinar.jpg',
+      },
+      {
+        Name: 'Learus Vacuum Cleaner',
+        Image: 'https://source.unsplash.com/featured/?zommconference',
+      },
+    ],
+  },
+  {
+    Name: 'Coaching',
+    Caption: 'IDI Training!',
+    contentPosition: 'right',
+    Items: [
+      {
+        Name: 'Consultation',
+        Image: 'https://source.unsplash.com/featured/?studying',
+      },
+      {
+        Name: 'Training',
+        Image: 'https://source.unsplash.com/featured/?students',
+      },
+    ],
+  },
+  {
+    Name: 'Speaking Engagement',
+    Caption: 'IDI Training!',
+    contentPosition: 'right',
+    Items: [
+      {
+        Name: 'Consultation',
+        Image: 'https://source.unsplash.com/featured/?studying',
+      },
+      {
+        Name: 'Training',
+        Image: 'https://source.unsplash.com/featured/?students',
+      },
+    ],
+  },
+  {
+    Name: 'Intercultural Competency Assessments',
+    Caption: '1:1 or Group Testing',
+    contentPosition: 'right',
+    Items: [
+      {
+        Name: 'Consultation',
+        Image: 'https://source.unsplash.com/featured/?studying',
+      },
+      {
+        Name: 'Training',
+        Image: 'https://source.unsplash.com/featured/?students',
+      },
+    ],
+  },
+]
